requirements.txt
django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
django_on_heroku==1.1.2
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.9.0
